[fx] fix fx self-test in zsh
fx self-test executes the default shell instead of bash (as defined in the
shebang), when calling each individual tests. This causes errors when
the default shell is zsh, since the tests are only compatible with bash.

This CL fixes it by calling bash instead of ${SHELL} when spawning the
test environment.

Change-Id: I8626ae4b20c26ea30d1f19aab46d1b47f07f7627