[swd] Replace many '_' with '-'
'_' is not valid in package names, so this change renames all packages
under //src/sys/pkg to use '-' instead of '_'.
For consistency across tests in this area, this change also renames many
binaries and cmx files to use '-' instead of '_'.
Finally, this change updates references to use the new GN target names.

Test: build, run tests, ota.  Ensure system boots and CLI tools work.
Change-Id: I044101e1eb98bf0b56da20955c0872895571245f