[bindc] Add disable_autobind option
The ability to disable autobinding is currently implemented as a bind
instruction. Eventually we want it to move to part of the component's
configuration (fxb/43400), but until then we need a way for drivers
using the new bind compiler to indicate that they shouldn't autobind.
This change implements it as an argument to the bind rule GN target.

Change-Id: Ia0d6dce2b0b155f8ba8898ece3f45b4cb4921179