[components] Avoid hub panicking when seeing a runner cap routed.
Currently, the hub code assumes that when a capability is being routed
to a component, that component must already be running. This isn't true
for runner capabilities, which are routed prior to the component
actually starting up, causing a panic when such an event is observed by
the hub.

Update the hub code to avoid tracking runners or storage (which
similarly should not appear on the hub).

Change-Id: I33ed204d3d14c2da61574fff5843180a05377c6b