[bindc] Generate output for bind program debugger.
For each statement in the bind program, information is printed about
whether the statement succeeded or failed. The debugger generates this
output as it executes the symbolic instructions, using the AST location
of each instruction to figure out which statement it was compiled from,
and produce the appropriate output for that statement type.

Change-Id: Iaecc1c39a930c6e431a083d05f80558f36bdb9f9