[fidl_codec] remove values' type field
Now values are not typed anymore. Some values like StructValue still
have a reference to a Struct but not a Type.

That means one less pointer in each value.

That also means that, for each function which needs the type (like
PrettyPrint and visitors), we have to provide the Type.

That means that you don't provide the type when you are creating values
but only when you are using them.

Change-Id: If6119513956c3e5153804c0509c3ade46dbe2fd6