[identity] Implement Oauth protocol on dev_auth_provider
This largely copies over the functionality already
in the AuthProvider protocol implementation. This
commit does not update token manager to use the
new protocol yet; token manager will be updated
after Google auth provider is also updated with
an Oauth implementation.

This change adds a serialization/deserialization
capability to access tokens, which enables tracking
things such as the fake account a token is issued
for.  This is necessary for ensuring consistency
between the user_profile_id reported by Authorize
calls in the token manager integration tests.
When using the old AuthProvider interface, this
consistency is achieved automatically because
only the GetPersistentCredential call needs to
report an account id.

Testing: unit tests

Bug: 565
Change-Id: I22bcc7e1df518c7e0e7f539807834ed7b6edc924