[scenic][a11y] Scenic Annotation API
This API allows clients with a ViewRef and a ViewHolderToken to
request Scenic to create a ViewHolder to hold an annotation View
that "draws on top of" (it inhabits the same View-space) the View
referred to by ViewRef.

a11y_mgr intends to connect to this service and use it to draw
highlights in tandem with the screen-reader.

Bug: 24332

Change-Id: Ib7f7ce5b2ff065641f7ceff20bfcdb1f400a29c9