[fidl][cpp] Read both hashed and explicit ordinals
HLCPP takes an approach similar to the other bindings,
which is to use a ::get_write_ordinal static method
that will convert either set of ordinals (hashed/explicit)
to the "ordinal" value specified in the IR when decoding.
This ensures that the internal tag_ value is always
consistent with the values inside the Tag enum, and that
the "ordinal" value is used during encoding without any
additional changes.

LLCPP is handled slightly differently because of the way
it decodes messages: since LLCPP reinterpret_casts the
raw message data into the generated xunion class, modifying
the internal ordinal_ value to take the "canonical" value
is not as simple as in other bindings. Thus, ordinal_ may
be either hashed or explicit, and the value used when encoding
is the value received when decoding (instead of always using
the "canonical" ordinal value). This behavior is documented
in xunion_test.cc.

Test: fx run-test fidl_tests
Test: fx run-test fidl_llcpp_types_test
Test: fx run-test fidl_llcpp_conformance_test
Change-Id: Ieb1b27178a8ba69795f8511734d7bfb30df62294