[wlan][mlme] Use Rust chan sched and scanner
Remove C++ channel scheduler and scanner. Use Rust version.

I had to change probe request to not include HT element since the hidden
AP I tested with would not respond. Our existing C++ implementation
also only included SSID and rates elements. Note that with HW scan
enabled in ath10k, the driver only include SSID.

See test section for verified functionalities. The primary thing that
does not work is that if we enable SW scan on ath10k, power state frame
is not sent when switching channel while client is connected. It is sent
on ralink, so doesn't seem to be implementation issue. This issue is not
a blocker since for ath10k, we use HW scan offload anyway.

Bug: 29063
Test:
1. Manual verification:
  a. HW scan works on ath10k.
  b. SW scan works on ath10k and ralink.
  c. After connecting, client remains connected.
  d. Scanning while client is connected:
    - Client remains connected
    - In SW scan, power state frame when going off and on channel on
      ralink.
  e. Hidden BSS is discovered when connecting with active scan setting.
2. Unit tests:
  a. fx run-test wlan_tests (this should cover all changes here except
     for sending power state frames when switching channel)
  b. Added unit tests in Rust: `fx run-test wlan-mlme-tests`

Change-Id: If46207ab11ded71ab179acaf6728ed90bbe2751b