[blobfs] Fix off-by-one in max_filename_size
When blobfs is queried for FilesystemInfo
(fuchsia.io.DirectoryAdmin:QueryFilesystem()), it reports a max
filesystem size of 65, which is off by one since blobfs files are
fact 64 bytes (the stringified length of a sha256 hash).

Fixed this off-by-one and changed the definition of kSha256HexLength to
*not* include the null terminator character (since we often use
fbl::StringView for filenames, which are explicitly not storing
C-strings).

Tested: Existing tests tweaked so expected query result is explicit
Change-Id: If58c6eb9fa4d4a455943034453ab55f95d1d9f1f