[devshell][contrib][fidl] Add tags generator for fidl files
This uses the IR .json files that fidlc generates during compilation for
communicating between fidlc and backends. Having a tags file allows
easier navigation in Vim (and probably will work in Emacs and others?)
to consts, enums, methods, interfaces, tables, etc. that are defined in
fidl files.

(It looks like a long script, but the majority of it is just test data.)

Test: scripts/generate-fidl-tags.py test
Change-Id: Idf000de2d5316db43f7af1657d192a46feae8957