[nmh][network_manager] Fixes invalid LAN creation.
Trying to create an interface with ports already in use fails as
expected. When it is tried a second time, is succeeds. This is because
when it was initially tried, the ports were incorrectly marked as
available. Only the temporarily reserved ports should have been made
available after a failure.

Test: Added test cases to reproduce problem.
      Verified they pass after the fix.

Bug: 43588

Change-Id: I4fbcdc2baf3d98020efc96c30c82bfa6adb15a59