[fuchsia.posix.socket] add {Base,Datagram,Stream}Socket
fuchsia.posix.socket.Provider/NewSocket implements a FIDL protocol that
provides a FIDL-based transport for datagram sockets. Add an
implementation of this protocol everywhere that
fuchsia.posix.socket.Control is supported.

The fdio function `socket` is not yet migrated to this new protocol.

This change is tested in two parts:
- the refactored implementation of fuchsia.posix.socket.Socket is
covered by existing tests in this CL's CQ run.
- the new implementation of fuchsia.posix.socket.Socket2 is covered by
existing tests in a child CL which switches fdio's `socket` to use the
new interface.

Bug: 21123
Change-Id: I56bda2e1f13b914d96d3c6c326f8bd51ff334acb