[nand-util] Generate wear histogram instead of just boundaries.
Now, using --wear will print out a histogram of the wear count for all
blocks, and log all values that are outside the expected range.

Test: Run the tool against a real device and see a histogram.

This CL basically increases the log verbosity of the tool (logging
being generally exempt from dedicated tests). That said, the main reason
for not having dedicated tests with this CL is that it would require
breaking the normal flow of the code just to add something observable
by a test, and that is likely to make the code harder to follow, for
very little gain.

Fixed: 39265

Change-Id: I32585cc05f6c5895e2a1dec60be9f3f8b99185cf