[acpica] Work around use-after-free
This implements a work around to the use-after-free present in ACPICA
when running on the NUC. The underlying problem is described in the
github isue (https://github.com/acpica/acpica/issues/416). The
workaround done here is to first delete the problematic objects, and
then delete the rest of the namespace subtree.

Test: Can pave+mexec on NUC with ASAN enabled.

Bug: 32590

Change-Id: Ia1587dd1ceb692142f9858544bb279b8a17757fa