[minfs] Validate shrink size
The math to compute size to grow to and the size
to shrink to for indirect vmo are asymmetric.
This is leading to some assertion failures.

This patch introduces a check on shrink size.

Once I get traces, I plan to replace the shrink math
with tested way to compute indirect vmo size
(VnodeBlockOffsetToIndirectVmoSize).

This patch is to gather more information on a race
that we are seeing in linked bug. We do not have a
reproducer for the race.

Bug: 41601

TEST: This adds logging under certain cases. I
      could not see the message getting printed
      as this bug is not yet reproducible at will.

Change-Id: Ic5224e55d3d58c5c95495571c48dedcecdd3f1de