[zircon][interrupts] Add INTERRUPT_UNMASK_PREWAIT_UNLOCKED
In the current design of the InterruptDispatcher it calls a
virtual method MaskInterrupt() while holding a spinlock. This
presents a problem in instances where a derived InterruptDispatcher
has code that needs to acquire a Mutex down the line. This change
introduces a new flag 'INTERRUPT_UNMASK_PREWAIT_UNLOCKED' which
indicates to WaitForInterrupt() that UnmaskInterrupt() be called
after releasing the dispatcher's spinlock. The derived dispatcher
is responsible for ensuring its own coherency when MaskInterrupt()
is called.

Test: Ran system interrupt tests, and confirmed with lockdep that
the locking issues experienced no longer occur.

Bug: 34434, 42282

Change-Id: Ib8021033d2e9d945c4565548786499ef64aab386