[carnelian] Improve touch sampling in infinite scroll example.
Devices with report rate at 60 Hz or lower require the
previous touch points to be saved as there's no guarantee
that we receive a new report with each update.

This also stops adjusting the sampling offset when too low
and just prints a warning. That's the preferred behavior in
a real app and still useful for determining what an ideal
offset is on a specific device.

Test: run infinite_scroll_rs
Change-Id: Ic408ace2ddcd359d2d3ae879d110d7f87c29b54c