[componentmgr] Routing follows refs
With this change, routing follows Realm references directly, instead of
looking up each component separately by absolute moniker. We also change
the signature of routing methods to take an `&Arc<Realm>` as the target
realm rather than a moniker, requiring the call site to perform a lookup
if it doesn't already have access to the Realm.

This has the advantage of making routing code less dependent on global
state, reduces the number of lookups, and supports future evolutions to
make the routing algorithm more modular.

This patch also includes the following changes, which were found to be
desirable in the course of the refactor:

- Previously, an action may continue holding on to an Arc<Realm> after
the action marked complete. This isn't a significant problem, but it
makes the change hard to test because it keeps the weak pointer valid.
To fix this, change `ActionSet::finish` drop the Arc<Realm> before
sending the notification. We also refactor ActionSet slightly to merge
the `register_action` and `finish_action` functions from `Realm` with
`ActionSet::register` and `ActionSet::finish`. This simplifies the
interface at the cost of more tightly coupling `Realm` and `ActionSet`,
which isn't a big deal since realms and actions are effectively coupled
already.

- Previously, `RoutingTest::destroy_dynamic_child` used breakpoints to
wait for the destruction to complete. Change this to call
`Realm::remove_dynamic_child` and wait for the action's notification;
this is simpler and ensures that the Arc<Realm> is dropped when the
function returns. This also requires a small change to
`remove_dynamic_child` to return the notification.

- Change the `Arc<Realm>` held by the `sys2.Realm` service
implementation to `Weak<Realm>`, and upgrade it between requests. This
protects against the possibility that a Realm is kept alive because a
sys.Realm service is still running.

Change-Id: I689d79055f0f95f636549de88feb0f2e47ed2fd0