[magma] Add manifest for msd_vsl_gc_nonhardware_tests
Should address the flaky test because the manifest contains
fuchsia.logger.LogSink.  The test failure has not been reproduced locally,
so this is a speculative fix; however this change removes the following
warning observed in the log just before the assert:

[584.084] 07415.07417> [WARNING:src/sys/appmgr/service_provider_dir_impl.cc(120)] Component fuchsia-pkg://fuchsia.com/magma_vsl_gc_nonhardware_tests#meta/msd_vsl_gc_nonhardware_tests.cmx is not allowed to connect to fuchsia.logger.LogSink because
[584.084] 07415.07417> this service is not present in the component's sandbox.

The package has also been renamed to msd_vsl_gc_nonhardware_tests for
consistency with other drivers.

Bug:43600

Change-Id: Iabda383c170bd2571a7d2fb221090609f3ba6e1f