[zircon][fit] Disable two StringView tests when assert() is not defined
StringView uses assert() to enforce some invarants.  Tests
AtThrowsExceptionWhenIndexIsOOR and CopyThrowsExceptionOnOOR expect
an assert() to fire and the process to be terminated.  That will only
happen when assert() is defined.  Disable these tests when assert() is
not defined.

This change is part of an effort to ensure the system works when GN
assert_level is 0.

Bug: 42977
Change-Id: Id759166d394b220b9b4a2f1daa3be0d40ea6c5af