[x86][platform-bus] Make x86 bus driver testable
This commit mostly refactors the X86 device class to make iteasier to
write tests for. The functional change to the class is to add a
destructor to cleanup the ACPICA library initialization that occurs
during the driver init process.

A simple device unit tests is added to exercise the ACPICA init and
shutdown process.

Test: Ran on NUC with driver.tests.enable=true added to the cmdline
and validated in the serial output that the tests ran and passed.

Change-Id: I1766413f9e316d58afdd3034259e04fc4b3e4f6c