[audio_record] Test fix.
Looks like the timeout call for audio tool allows recording for a few
more milliseconds. Hence adjusting the file size check to not check for
an exact number and only check if the file size is not less than
1.5seconds when asked to record 2 seconds. This is better as this could
reduce flake and also the original purpose of the test is to check audio
quality and not if timeout triggers at a particular point.

Bug: fxb/43480
Test: Verified locally that the test passes now.
Change-Id: I47a6d32c3e085d34e402bfd11dc384c8c09c91e6