[tools][netboot] Use client port for queries.
Fixed: 43638

When issuing queries for device discovery, some clients are not able to
find devices, but are able to find them when using the netaddr binary.

When using wireshark it appeared the only major discrepancy in network
traffic was that the src port was randomly chosen when using this
library (tools/net/netboot), whereas the netaddr binary was using a
deterministic port.

This fixes the use of the protocol by adhering to the implementation for
port-binding visible here:
https://fuchsia.googlesource.com/fuchsia/+/0e30059/zircon/tools/netprotocol/netprotocol.c#59

Which is implemented to adhere to ufw firewall rules defined here (also
fixed in this CL (see below)):
https://fuchsia.googlesource.com/fuchsia/+/0e30059/tools/devshell/setup-ufw#55

This also fixes the UFW setup rules to include the ports that the
netboot code is attempting to bind to.

Test: Verification via wireshark. CQ verification (tested in botanist).
Change-Id: I31c3ddf4ad96d3af093818a0c7c6b0ac821589c1