[usb] xHCI rewrite
This patch contains a full re-write of the xHCI driver.
The new driver has been written based on the documentation
available at https://www.intel.com/content/dam/www/public/us/en/documents/technical-specifications/extensible-host-controler-interface-usb-xhci.pdf.

The objective of this rewrite is to provide a fast, compliant,
and more testable version of xHCI than its predecessor.
This CL only contains the implementation of xHCI, not the tests.
The tests will be submitted in a future CL.

This new driver provides the following improvements:

* Fully asynchronous state machine driven directly by IRQs from hardware.
This asynchronous model provides improved traceability and performance in the system
without the need for blocking calls, including during device setup.
This can be used to significantly improve the parallelism of the system, without
the need for additional context switching. Each callback corresponds directly
to a completion event placed on the event ring.
* Significantly improved performance compared to the old driver.
* Full conversion to C++ and ddktl. Each class in the driver roughly corresponds
to a hardware block as defined in the xHCI specification. These abstractions are
designed to be as close to hardware as possible, and can easily be traced back to
the relevant section of the xHCI specification.
* Heuristics to switch between a new "real-time" polling mode
to achieve high-throughput for gigabit speed transfers on supported hardware, and
an interrupt-driven mode for interrupt-driven transfers.
* Add proper support for USB 2.0 hubs.

Test: https://fuchsia-review.googlesource.com/c/fuchsia/+/345856

Change-Id: Ie8574d30d4ceb81b9ece88ea613039b27820697e