[netstack2] Receive on test notify channel in tests before checking empty queue
Wait for a signal from the test notification channel in NDP tests before
checking if the events queue is empty. This is to resolve a race condition where
a test will check the event queue if it is empty after the last event has been
popped from the event queue but has not been handled. Since the signal will only
be sent after handling the last event in the queue, we know that if the queue is
empty after receiving the signal, then all the events in the queue have been
handled.

Fixed: 43599

Test: ran the command below without errors:
`fx run-test netstack_gotests -t netstack_test -- --test.run TestNDPIPv6RouterDiscovery --test.count 10000 --test.failfast`

Change-Id: Ic45e43fb680016448b7a6704fe11503b92b9ee7e