[system-updater] require ZBI in the update package
If we have an device running an old version of Fuchsia, it's possible
that the bootloader in a bootloader-only package doesn't support the
old kernel. Instead this patch changes the system updater to require
that update packages always contains at least a ZBI to make sure we are
always updating the device to the latest version.

Fixed: 43598

Change-Id: Ib9945c3ce08f78bd78342658188960f506343388