[fx][contrib] Add `fx roll-compiler` command
This is a handy script to do either of the two steps of rolling a
out a new compiler built by Fuchsia's toolchain bots:
1. `fx roll-compiler goma` finds the latest (or specified)
   CIPD packages and publishes them to Goma.
2. `fx roll-compiler fuchsia` finds the Goma-published (or specified)
   CIPD packages and updates the prebuilts Jiri manifest file.

By default this does the right thing for Clang.
With `--package fuchsia/third_party/gcc` it should handle GCC.
With `--package fuchsia/rust` and maybe --platforms it should handle Rust.

Change-Id: Idf6a84712536bfcda5508f20a066dd19cd0d4501