[zxio] Align zxio_node_attr_t towards fuchsia.io2
To ease the addition of io2 backends, this CL aligns the "node attribute"
type closer to io2 as opposed to io1. Each element in the node attribute
table in io2 is optional; I think it is desirable to expose this
property in the zxio public interface, and have fdio provide fallback
values.

The next CL will be building on top of this to add a io2 backend.

Change-Id: Ic6ad09d51adee4266a32fe8cf6145ed9d662334b