[debugger] Console improvements.
Stack formatting options are separated into their own struct and the
FormatFrameList() is renamed FormatStack(). The "simple" format function
is removed because it is unused. This part should have no
user-observable change.

Changes unsymbolized "verbose" stacks to always print out the address
first, even if the "IP" id displayed below. It looks funny with no
information on the first line.

Fix a use-after-free I noticed in the "up" command because it captures
the command which does not have owning pointers to the thread/target it
uses.

Computes the new frame index from within the "up" command callback in
case it is executed in succession faster than the async IPC can respond.

Elides more information from the "up" and "down" commands. Previously it
could output more than one screen of information for some Rust frames.

Adds a test for the "up" command. As part of this I noticed that the
MockConsole returns the output in first-in-last-out which seems to be a
bug. The order is reversed and several affected tests are updated.
Mostly the ordering was masking some assumptions that Clear() was the
same as FlushOutputEvents() (Clear() clears the screen, not the output
event queue).

Change-Id: I9372a33e12a2442db69d57fc4460a26f7a8cdac8