[sfw][input] Send modifiers in KeyboardEventDescriptor
Currently, no duplicate InputEvents are sent for held keys. This means
held modifier keys are not propagated after their first appearance in an
InputReport.

This change adds modifiers as a separate field in the
KeyboardEventDescriptor so clients can still expect a single
`KeyEventPhase::Pressed` event for modifiers, but not have to track the
state of modifiers separately.

TEST= fx run-test input_tests

Change-Id: I8d526a7b358204444b35c5dbc29e88cebf4b3596