[debugger] Add attach-job verb.
This new verb takes the job attach logic from "attach".

Previously, attach had three jobs: creating a filter, attaching to a
process, and attaching to a job. But the string "job 2 attach X" is
ambiguous about whether you're creating a filter for the given job, or
trying to attach the given job.

The separate job-attach verb removes this ambiguity. Having a longer
command should not be too bad since this is an uncommon operation.

Some additional features over the old implementation:
  - A new job context is automatically created if the current one is
    attached.
  - The job-attach can now take a list of filters to apply to that job
    which makes the task of "get processes for job X" one step.

Removed the "Clone" stuff from job contexts. It was not used.

BUG: 43558, 43556
Change-Id: Ic9933a1283e2b973c7c06fd6ff277f68ebe7e521