[debugger] Remove Setting struct.
I found the multiple layers between a Setting, a SettingValue, and
SettingInfo confusing and this makes it more difficult to implement
SettingsStore which we are considering virtualizing.

Removes Setting and SettingInfo. The data is stored directly in the
information stored in the schema which was renamed from SchemaSetting
(avoiding confusion with the SettingSchema class it's inside of) to
Record.

Simplified the format_setting tests which no longer need to have a
SettingStore to test setting formatting.

Change-Id: I6d4105dc08bfc7d3d252f2bff9c3cc02443c5aa2