[debugger] Improve filter listing messaging.
There was a bug where "filter" would display the filters for the current
job only. This bug is now fixed.

Improves the messaging to make explicit when a subset of filters are
displayed (inside the current job).

Implement the WillDestroyFilter() event in the ConsoleContext.
Previously filters were not cleaned up which would cause use-after-free
errors.

Change-Id: If2bb9826c3ba9bbc1436b0e3667fea1aa5b6cc4d