[audio_core] Init source links with object
This CL changes source link initialization to use the source object and
its stream instead of the link.

Change-Id: I40bc5b650b4d31dd78bfb8c93a45ca394b1e161a

---

This is part of a series of changes which remove AudioLink from AudioCore.
AudioLink requires an implicit distributed ownership model where
destination objects own sources, and sources own destinations, requiring us
to manually free objects.

The new feature pressuring this change is the addition of a
LoudnessTransform to links, which must be added to links by the RouteGraph.
The transform determines what loudness stages take effect for a given
output.

The replacement is the LinkMatrix, a single owner for all AudioObject
connections. RouteGraph will own the LinkMatrix and all AudioObjects.

Initially this refactor will not be strict about ownership, preferring
shared_ptrs.

Once some restructuring in audio core is done to ensure drop order
consistency across threads, more strictness can be imposed at compile
time.

Bug: 39624
Bug: 39426
Bug: 43514

TEST: no behaviour change; fx run-test audio_core_unittests
Change-Id: I4a86b35280dca304df9bf71dd07389780ade0b99