[x86][platform-bus] Refactor IOMMU management
Turns the C-style IOMMU management code into more idiomatic C++ style
and allows it to be instantiated as an object. This makes it easier to
test and unit tests against the DMAR table parsing are added.

The existence of a global intstance of the manager is preserved for
the moment, but can be removed once the call sites have been migrated.

Test: Run on NUC  with "driver.tests.enable=true" and
"driver.iommu.enable=true"

Change-Id: I16b4702079772b80fa96f7332ee34069ff0d7f4a