[fidlc] Add validity checks to SourceElement
This change adds a new function to `SourceElement` to check whether a
`SourceElement` has a valid span. This may be useful before trying to
access `span()` which panics when no span exists.

Furthermore, this change also fixes a masked issue where an invalid
`SourceElement` access would cause the logic inside the assertion to
segfault instead of the assertion itself panicking. While both did
result in correctly crashing the compiler, the new backtrace on the
assert is now much easier to debug.

Bug: 8027
Bug: 43740

Test: fidl-compiler-test
Change-Id: I960ca7b853680b5d243c7055103f3707cb2797ed