[fidl][go] with-ctx DispatchImpl and go interface
`DispatchImpl()` is intended to replace `DispatchImplWithCtx()`. The
marshaler context is now stored inside an implementor of the new Context
interface type (which is mostly identical to context.Context).  The
BindingKey is stored in the same way in the Context, which enables FIDL
handler implementations to correlate requests with the request sender
such that hanging gets can be implemented. A new Go interface is
generated per FIDL protocol where the only difference is that every
handler function has this Context as its first parameter. The equivalent
TransitionalBase has been generated per FIDL protocol as well.

Bug: 39618
Test: fx run-host-tests fidlgen_golang_test
      fx run-host-tests fidlgen_golang_ir_test
Change-Id: Id9bddb1835415de8b0012a24ab688ced358df649