[fidl][lint][syscalls] Make fidl-lint support CStyle for syscalls
In addition to the normal style (now called IpcStyle), autodetect when
running on syscalls, and switch to CStyle. This changes bits, protocols,
protocol methods, structs, tables, unions, xunions, and enums to be
lower_snake_case, rather than UpperCamelCase. All other casing is
unchanged between styles.

Currently, the decision to use CStyle is based on the library name.
Ideally this would probably be keyed off of Transports but that has some
complexities that we don't wish to deal with right now.

Bug: 43642
Test: new cases in lint_findings_tests
Change-Id: If5b24c9bcd3b74c073915be5a2f611e0844519d9