[component_manager] Support cloning null stdio when spawning processes
Cloning a null stdio handle should result in a null stdio handle for the
child process, rather than an error. We already do this for closed
stdio handles, but not null ones (e.g. if the current process was
launched without a stdin).

This changes the error handling for null handles to match closed ones.
It corresponds to a change merged to fdio: fxr/315686

Finally, change the Rust fdio bindings to accept any AsRawFd/IntoRawFd
type, not just File, when building SpawnActions. Technically this is
putting a pair of std::os::unix traits in our "public" API, which we
may want to swap out with Fuchsia-specific traits one day. I don't see
this as a problem, since (1) we don't have API stability for Rust and
(2) when we do stabilize, it'll be trivial to swap these traits out with
other traits.

This fixes a failure in component_manager_tests when rust_panic="abort".
Bug: 26968

Test: The only way to test this is to use the elf_loader in a process
with a null stdio handle. Since the root_job_test_runner already runs
the elf_loader tests in a custom environment, I modified it to also
provide a null handle for stdin, which is not needed by the test.

Change-Id: I08af2a2246ce04a449be5509c77a2f17d222303f