[audio_core] Define LoudnessTransform
This class will centralize some duplicated logic between AudioRenderer and
AudioCapturer.

An implementation of this interface will be attached to AudioLinks,
instead of the current VolumeCurve attachment.

This will also enable the RouteGraph to attach an implementation of this
interface which rejects all stages when routing streams to outputs with
independent volume control.

Bug: 43514
TEST: fx run-test audio_core_unittests
Change-Id: Iaa974795da02d8c02b3a105068ecf9281032f50c