[components][flake] StartEvent -> BeforeStartEvent
`on_route_capability_async` assumes that a component that is requesting
a capability must have an `Execution` populated in the hub's instance
map. This is not necessarily the case, because the Start hook is invoked
after the component is started.

To fix, turn StartEvent into BeforeStartEvent and invoke it just before
the instance is started, but after the Runtime is generated. Also, pass
a reference to the Runtime in the EventPayload; we can't have hub take a
lock on the ExecutionState anymore because it would deadlock.

The previous behavior was causing a race in hub. `hub/exec/used` is
populated when a capability is routed. Because StartHook is invoked
after the component is started, it's possible for the component to use a
capability before the Runtime state (and the hub's exec directory) is
populated. When hub encountered this scenario, it panicked.

Also, change the implementation of hub not to panic, to defend against
runners that could continue executing code after the component is
"stopped".

The previous behavior caused storage_integration_test to flake due to
the panic in hub, but can likely affect other integration tests as well.

BUG=43210

Change-Id: I09251afe314733f5d3ebdd8caadc204c840bf42e