[debugger] Add pretty frame matching system.
It has the ability to match sequences of stack frames using a very
simple wildcard scheme. Frames can be matched by exact function name, by
filename from the right, or by wildcard.

It is currently not hooked up to anything yet so there will be no
behavior change.

The MockStackDelegate is pulled out of stack_unittest.cc to a separate
header so it can be re-used in the new unit tests.

The code that matches file names from the right on path boundaries is
moved out of the index implementation to a helper in file_util so it can
be shared with the matching code.

An include is added to the symbol_data_provider. Using the
SymbolDataProvider requires this header so it's better for callers for
it to be automatically included.

Change-Id: Iff5d33f04fc5d745432aecd43f904c65e9f137c9