[i18n] Move rust_icu to third_party/rust_crates.
This change removes the need to manage the library in two places.  The
library is now hosted at https://github.com/google/rust_icu.

The change modifies all dependencies in the tree to make it an atomic
update.  Dependencies used are `log` and `paste`, both of which are
already used in Fuchsia's code base.

OSRB approval is at: fxb/43526.

Bug: 43554
Change-Id: Idff02560a611afd5de49abc715c83617ea81101b