[fidl][llcpp] Fix sending empty tables
We would expect that the
https://bugs.fuchsia.dev/p/fuchsia/issues/detail?id=40782#c6 idea of
using some valid pointer value in the table vector header would work,
since it appeases the linearizer. However, it turned out that
empty tables don't go through linearization, going straight to the
encoder which would fail. They don't get linearized because their
max-out-of-line value as inferred by the compiler is zero, because
the maximum number of elements is zero. LLCPP then assumes that
this structure doesn't have pointers, and don't apply
linearization.

We could add another type shape visitor that outputs if the
message has a pointer, and use that to determine running
linearization. I found the message depth property nicely describes
if the message has a pointer. This CL switches the
linearization criterion to that, and have the linearizer handle
the absent vector case.

Fixed: 40782

Change-Id: Iff5c8131c3636fb1628229d3bf31eb0cb2533802