[testsharder] Make shard names more consistent across builds
Changes to the build graph that affect tests.json can result in
confusing shard renamings – e.g., QEMU-(1) in this build [0] corresponds
more closely to QEMU-(3) in the next [1].

0: https://ci.chromium.org/p/fuchsia/builders/try/fuchsia-x64-asan/b8892238158657457424
1: https://ci.chromium.org/p/fuchsia/builders/try/fuchsia-x64-asan/b8892238525732415280

We can make this less confusing by sorting the shards by the name of the
first test before assigning them names based on their ordering. As long
as the change to the build graph doesn't affect a test that's very early
in a shard, this will ensure that given shard name corresponds to
roughly the same set of tests (or at least the same set of long-running
tests, since those come first in the shard) across many builds.

Change-Id: Id3fdace6ff87422c03a7fcef4690c2ba25dc5597