[fdio] Avoid allocating zero-sized VLAs and re-enable UBSan
The fdio library uses variable length arrays to translate to zxio types
for the sendmsg and recvmsg operations. Since variable length arrays
must have nonzero sizes, we have to handle the empty vector case
differently from the nonzero case and pass null with a zero length.

This also re-enables UBSan for fdio, which caught this bug in sendmsg
and catches the same bug in recvmsg with the newly added test case.

Tested by fdio-test under UBSan

Bug: 43773
Change-Id: Ib1fb8c36e10ec2e058023fce51f1fb3a6b0c6ede