[devshell] Raise default symbolizer restart interval to 15.
Increases the number of stack frames symbolized per invocation of
llvm-symbolizer. Chromium stack traces, which tend to be fairly deep
and can easily have up to 30 or 40 frames per thread, can now be
symbolized in a few seconds.

The existing limit can take 30s to over a minute to symbolize a Chromium
crash dump.

Bug: 42593
Test: tested on a Chromium CHECK-failure in Blink.
Change-Id: I94776f68cc4efc51c8709bfe955bb16ac3bb0d97