[escher] Test doesn't fail on Vulkan Performance Warnings.
Fixed: 36827

Sometimes we use sub-optimal image formats / layouts / render
configurations for simplicity and in that case there may be
some performance warnings. These usages don't violate the vulkan
specs and it is okay to keep them in tests.

After discussion with jjosh@, we decided to not fail the test
if there are performance errors, but stil print out logs instead.

TEST=escher_tests

Change-Id: I4cadf0b01ae1667ec5b80383af56bc12a037a42e