[components] Scoped Breakpoint System
This CL introduces the notion of a ScopedBreakpointSystem. The BreakpointSystem
now vends ScopedBreakpointSystem objects which have an attached AbsoluteMoniker.
The AbsoluteMoniker represents the realm to which breakpoints will be received.
If a breakpoint invocation has a target_moniker within the realm of the provided
AbsoluteMoniker, then it will dispatched to the corresponding
ScopedBreakpointSystem.

All ScopedBreakpointSystem objects share a BreakpointRegistry. However,
now InvocationSenders in the registry have a scope (an AbsoluteMoniker).
When it comes time to dispatch an event to breakpoint systems, the event goes to
all InvocationSenders but may be rejected by certain InvocationSenders that are
not a realm of the event. A new operation on AbsoluteMoniker "is_realm_of" is
introduced to decide whether an InvocationSender can send a particular event.
In a nutshell, an AbsoluteMoniker is a realm of another AbsoluteMoniker if its
path is a prefix of the other moniker.

This CL also renames the interposer directory in tests to breakpoint_system and
adds a second breakpoint system test: echo.

This test verifies that a scoped breakpointsystem cannot see capabilities that
it does not dispense or use directly.

echo_reporter is the parent of echo_server
echo_client is a sibling of echo_reporter.

echo_client uses the echo service from echo_server. However echo_reporter does
not intercept this.

echo_reporter uses a scoped breakpoint system. This test verifies that
echo_reporter cannot see echo_client's connection to echo_server. echo_reporter
uses an EventSink of RouteCapability that it then echos out to the integration
test.

This required echo_reporter to implement trigger, so that it knows when to start
soaking events and when to drain events.

This also required the integration test to implement the echo service in a
clean, modular way.

Both trigger and echo where modularized into echo_capability.rs and
trigger_capability.rs that can be used in other tests.

Change-Id: I1527fd78891ee760584e62122d92b993af7f8fed